requirements.txt
aiogram==2.21
aiohttp==3.8.1
aiosignal==1.2.0
amqp==5.1.1
async-timeout==4.0.2
attrs==21.4.0
Babel==2.9.1
billiard==3.6.4.0
celery==5.2.7
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
dnspython==2.3.0
flower==1.2.0
frozenlist==1.3.0
greenlet==1.1.2
humanize==4.6.0
idna==3.3
importlib-metadata==4.12.0
importlib-resources==5.9.0
kombu==5.2.4
Mako==1.2.1
MarkupSafe==2.1.1
multidict==6.0.2
prometheus-client==0.16.0
prompt-toolkit==3.0.36
pymongo==4.3.3
python-decouple==3.6
pytz==2022.1
six==1.16.0
tornado==6.2
vine==5.0.0
wcwidth==0.2.6
yarl==1.7.2
zipp==3.8.1
